requirements.txt
tensorflow>=2.16
numpy
pillow
scikit-learn
matplotlib
streamlit
os
